Implement transaction delete route

+++ src/database/migrations/1593354526901-CreateTransaction.ts
@@ -25,7 +25,7 @@ export default class CreateTransaction1593344537508
           },
           {
             name: 'value',
-            type: 'numeric',
+            type: 'float',
           },
           {
             name: 'type',
